Refactor into micro-analysis
Before this commit there was 1 single project containing
a proof-of-concept for both lttng-ust and lttng-kernel traces.

After this commit, we break the big single project into
smaller and more specific projects.
Each of the new projects is a "contra microanalysis", i.e.
a containerized analysis for a specific trace type.