Create LICENSE
I think is possible to release as MIT, see below:

The EPL isn't a copyleft license. The EPL license doesn't claim to apply to program modules that are linked to the EPL code.

You can choose any compatible license you like for your own code.
The Apache and MIT licenses are both compatible with the EPL, but the GPL isn't.